langchain/requirements.txt
requests
aiohttp
asyncio
psycopg2-binary
python-dotenv
supabase
geopandas
pandas
numpy
shapely